Enhancement!
 - Endrer navn på filer
 - Lager en compact BloggRollItem
 - Gjør søkefelt finere i Arkivet
 - Nå søker kunn etter dato og tittel
 - Fjerner styling i bloggRoll som ikke gjorde noe
