+++ src/components/BlogRollItem.jsx
@@ -5,7 +5,7 @@ import { Box, Heading, Text, Paragraph } from "grommet"
 
 import PreviewCompatibleImage from "./PreviewCompatibleImage"
 
-const BlogRollItem = ({ image, slug, title, description, date }) => (
+const BlogRollItem = ({ image, slug, title, description, date, compact }) => (
   <Box
     elevation="xsmall"
     animation={{ type: "fadeIn", size: "medium" }}
@@ -14,7 +14,7 @@ const BlogRollItem = ({ image, slug, title, description, date }) => (
     onClick={() => navigate(slug)}
   >
     <Box
-      height="small"
+      height={compact ? "xsmall" : "small"}
       background={{ dark: false, color: "light-2" }}
       round={{ corner: "top", size: "xsmall" }}
     >
@@ -26,10 +26,19 @@ const BlogRollItem = ({ image, slug, title, description, date }) => (
       />
     </Box>
     <Box margin="medium" flex>
-        <Heading color="black" level="2" size="small" margin="none">
+      {compact 
+      ? <Heading color="black" level="3" size="small" margin="none">
           {title}
-        </Heading>
-      <Paragraph margin={{ vertical: "xsmall" }}>{description}</Paragraph>
+        </Heading> 
+      : <>
+          <Heading color="black" level="2" size="small" margin="none">
+            {title}
+          </Heading>
+          <Paragraph margin={{ vertical: "xsmall" }}>{description}</Paragraph>
+        </>
+      }
+      
+      
       <Box flex justify="between" direction="row" align="end">
         <div></div>
         <Text size="xsmall" color="grey">
