--- src/pages/arkivet.md
@@ -1,7 +1,0 @@
---- 
-templateKey: arkiv
-title: Arkivet
-image: /img/upload/sauejordet.jpg
-description: >-
-  Her finner du gamle artikkler/arangementer.
----
