+++ src/pages/artikkler.md
@@ -1,0 +1,7 @@
+--- 
+templateKey: Archive
+title: Alle artikkler
+image: /img/upload/sauejordet.jpg
+description: >-
+  Her finner du utdaterte artikkler.
+---
